refactor cloud storage actions into separate component
